Rename Footnote files as Math and restores Footnote functionality
